move themes back to public
needed for storybook to inject theme variables
